restrict user access

+++ praktikum09/Codeigniter-3/Databases-CRUD-Session/application/views/mahasiswa/index.php
@@ -31,9 +31,11 @@
                      <div class="col-12">
                          <h2>Daftar Mahasiswa STT Terpadu Nurul Fikri</h2>
                      </div>
-                     <div class="form-group row ml-1">
-                         <button name="submit" type="submit" class="btn btn-primary"><a href="create" class="text-white">Tambah Data Mahasiswa</a></button>
-                     </div>
+                     <?php if ($this->session->userdata('ROLE') == 'ADMIN') { ?>
+                         <div class="form-group row ml-1">
+                             <button name="submit" type="submit" class="btn btn-primary"><a href="create" class="text-white">Tambah Data Mahasiswa</a></button>
+                         </div>
+                     <?php } ?>
                      <div class="col-12 table-responsive">
                          <table class="table table-hover table-bordered">
                              <thead class="thead-dark align-center">
@@ -57,10 +59,12 @@
                                          <td style="vertical-align: middle;"><?= $mhs->prodi_kode ?></td>
                                          <td style="vertical-align: middle;">
                                              <a href="detail?id=<?= $mhs->nim ?>" class="pr-2"><i class="fa-solid fa-eye pr-1"></i>View</a>
-                                             <a href="edit?id=<?= $mhs->nim ?>" class="pr-2"><i class="fa-solid fa-file-pen pr-1"></i>Edit</a>
-                                             <a href="delete?id=<?= $mhs->nim ?>" class="pr-2" onclick="if(!confirm('Anda Yakin Menghapus Data Mahasiswa Bernama, <?= $mhs->nama ?> ?')) {return false}">
-                                                 <i class="fa-solid fa-trash pr-1"></i>Delete
-                                             </a>
+                                             <?php if ($this->session->userdata('ROLE') == 'ADMIN') { ?>
+                                                 <a href="edit?id=<?= $mhs->nim ?>" class="pr-2"><i class="fa-solid fa-file-pen pr-1"></i>Edit</a>
+                                                 <a href="delete?id=<?= $mhs->nim ?>" class="pr-2" onclick="if(!confirm('Anda Yakin Menghapus Data Mahasiswa Bernama, <?= $mhs->nama ?> ?')) {return false}">
+                                                     <i class="fa-solid fa-trash pr-1"></i>Delete
+                                                 </a>
+                                             <?php } ?>
                                          </td>
                                      </tr>
                                  <?php endforeach ?>
